+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+- Remove category validation on create & update group operations [#352](https://github.com/rokwire/groups-building-block/issues/352)
+
 ## [1.14.0] - 2023-01-30
 ### Added
 - Integrate govulncheck within the build process [#319](https://github.com/rokwire/groups-building-block/issues/319)